added potential logic for tests
