+++ build.gradle
@@ -118,6 +118,7 @@ dependencies {
     testImplementation "org.testcontainers:junit-jupiter:$tcVersion"
 
     testImplementation group: 'ch.qos.logback', name: 'logback-classic', version: '1.2.13'
+    testImplementation group: 'org.mockito', name: 'mockito-core', version: '4.8.0'
 
     testImplementation group: 'commons-io', name: 'commons-io', version: '2.11.0'
 }
