+++ src/test/java/coop/f1r3fly/fs/struct/Utils.java
@@ -1,6 +1,13 @@
 package coop.f1r3fly.fs.struct;
 
+import org.apache.commons.io.FileUtils;
+
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.util.HashMap;
+import java.util.List;
 
 public class Utils {
 
@@ -24,6 +31,37 @@ public class Utils {
 
         public static <L, R> Pair<L, R> pair(L left, R right) {
             return new Pair<>(left, right);
+        }
+    }
+
+    public static void cleanDataDirectory(String destination, List<String> excludeList) {
+        try {
+            // if test fails, try to cleanup the data folder of the node manually
+            // cd data && rm -rf blockstorage dagstorage eval rspace casperbuffer deploystorage rnode.log && cd
+            cleanDirectoryExcept(destination, excludeList);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+    public static void cleanDirectoryExcept(String directoryPath, List<String> excludeList) throws IOException {
+        File directory = new File(directoryPath);
+        Path dirPath = directory.toPath();
+        File[] files = directory.listFiles();
+
+        if (files != null) {
+            for (File file : files) {
+                Path filePath = file.toPath();
+                String relativePath = dirPath.relativize(filePath).toString();
+
+                if (!excludeList.contains(relativePath)) {
+                    if (file.isDirectory()) {
+                        FileUtils.deleteDirectory(file);
+                    } else {
+                        Files.deleteIfExists(file.toPath());
+                    }
+                }
+            }
         }
     }
 }
